rationapi/requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.2
django-admin==1.3.2
django-allauth==0.39.1
django-cors-headers==3.2.1
django-excel-response2==2.0.8
django-filter==2.1.0
django-impersonate==1.4.1
django-rest-auth==0.9.5
django-six==1.0.4
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
keyring==10.6.0
keyrings.alt==3.0
mysqlclient==1.4.2.post1
oauthlib==3.0.2
Pillow==5.1.0
pipenv==2018.11.26
protobuf==3.0.0
psycopg2==2.8.3
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
PyPDF2==1.26.0
pypi==2.1
python3-openid==3.1.0
pytz==2019.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.2.0
screen==1.0.1
SecretStorage==2.3.1
simple-crypt==4.1.7
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.2
virtualenv-clone==0.5.3
webencodings==0.5.1
whitenoise==4.1.3
xhtml2pdf==0.2.4
xlwt==1.3.0

